requirements.txt
pycryptodome~=3.20.0
